vscode: add owner to task problem matchers
If previously a problem matcher matched a task output a problem was
generated and showen in vscode. But if the problem was fixed and the
task output no longer had the matching error output, the error was not
removed. As mentioned in github.com/microsoft/vscode/issues/129341,
setting a owner fixes this.