ci: adapt DooD aliases to GitHubActions
The GitHub Actions CI runs in a container itself and clones the
repository therein. So volume-mounting the path 1:1 does not work.
Instead we have to rely on the default Runner volume mount that is
visible in the GitHub Actions logs: `--volume /home/runner/work:/__w`.
Additionally we cant run interactively an/or with an attached tty.
Otherwise we get lots or warnings like "bash: cannot set terminal
process group (494): Inappropriate ioctl for device" and "bash: no job
control in this shell".