rtl/neorv32_wbp_io: fix fsm lockup
If a slave responded with asserting `err`, the rvsc FSM did lock up and
kept `cyc` asserted indefinitely. This fixes it by also delaying
back-to-back transaction where the first transaction was aborted with an
error. It also fixes the if/else ordering of `rvsc.pending` such that it
is correctly reset on slave errors.